bugfix: add quote to context
add if statement to only fetch quote if it doesn't have one already
